cli: implement `preview` subcommand

+++ src/types/configuration.rs
@@ -27,7 +27,7 @@ pub struct Service {
     pub autorestart: bool,
     pub namespace: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub port: Option<u16>,
+    pub port: Option<u32>,
     pub stdout: String,
     pub stderr: String,
     #[serde(skip_serializing_if = "Option::is_none")]
